+++ HISTORY.rst
@@ -12,6 +12,7 @@
 * Add database conversion test to the common app.
 * Fix label display for resolved smart links when not using a dynamic label.
 * Only show smart link resolution errors to the user with the smart link edit permission.
+* Intercept document list view exception and display them as an error message.
 
 3.1.1 (2018-09-18)
 ==================